Create BaseNavigatorView
