+++ tests/unit/views/DocumentationTopic.spec.js
@@ -11,55 +11,27 @@
 import * as dataUtils from 'docc-render/utils/data';
 import { shallowMount, RouterLinkStub } from '@vue/test-utils';
 import DocumentationTopic from 'docc-render/views/DocumentationTopic.vue';
-import DocumentationTopicStore from 'docc-render/stores/DocumentationTopicStore';
 import onPageLoadScrollToFragment from 'docc-render/mixins/onPageLoadScrollToFragment';
 import DocumentationNav from 'docc-render/components/DocumentationTopic/DocumentationNav.vue';
 import NavBase from 'docc-render/components/NavBase.vue';
-import AdjustableSidebarWidth from '@/components/AdjustableSidebarWidth.vue';
-import NavigatorDataProvider from '@/components/Navigator/NavigatorDataProvider.vue';
-import Language from '@/constants/Language';
-import Navigator from '@/components/Navigator.vue';
 import { TopicSectionsStyle } from '@/constants/TopicSectionsStyle';
-import { storage } from '@/utils/storage';
-import { BreakpointName } from 'docc-render/utils/breakpoints';
-import StaticContentWidth from 'docc-render/components/DocumentationTopic/StaticContentWidth.vue';
+import BaseNavigatorView from 'docc-render/views/BaseNavigatorView.vue';
 import onThisPageRegistrator from '@/mixins/onThisPageRegistrator';
 import { getSetting } from 'docc-render/utils/theme-settings';
 import { flushPromises } from '../../../test-utils';
 
 jest.mock('docc-render/mixins/onPageLoadScrollToFragment');
 jest.mock('docc-render/mixins/onThisPageRegistrator');
-jest.mock('docc-render/utils/FocusTrap');
-jest.mock('docc-render/utils/changeElementVOVisibility');
-jest.mock('docc-render/utils/scroll-lock');
-jest.mock('docc-render/utils/storage');
 jest.mock('docc-render/utils/theme-settings');
 
 const defaultLocale = 'en-US';
 
-const TechnologyWithChildren = {
-  path: '/documentation/foo',
-  children: [],
-};
-
-const navigatorReferences = { foo: {} };
-
-jest.spyOn(dataUtils, 'fetchIndexPathsData').mockResolvedValue({
-  interfaceLanguages: {
-    [Language.swift.key.url]: [TechnologyWithChildren, { path: 'another/technology' }],
-  },
-  references: navigatorReferences,
-});
 getSetting.mockReturnValue(false);
 const routeEnterMock = jest.spyOn(dataUtils, 'fetchDataForRouteEnter').mockResolvedValue();
 
 const {
-  CodeTheme,
-  Nav,
   Topic,
-  QuickNavigationModal,
 } = DocumentationTopic.components;
-const { NAVIGATOR_HIDDEN_ON_LARGE_KEY } = DocumentationTopic.constants;
 
 const rootLink = {
   path: '/documentation/technologies',
@@ -160,25 +132,15 @@ const schemaVersionWithSidebar = {
   patch: 0,
 };
 
-const AdjustableSidebarWidthSmallStub = {
-  render() {
-    return this.$scopedSlots.aside({
-      scrollLockID: AdjustableSidebarWidth.constants.SCROLL_LOCK_ID,
-      breakpoint: BreakpointName.small,
-    });
-  },
-};
-
 const stubs = {
-  AdjustableSidebarWidth,
-  NavigatorDataProvider,
+  Nav: DocumentationNav,
+  NavBase,
+  BaseNavigatorView,
+  'router-link': RouterLinkStub,
 };
-
-const provide = { isTargetIDE: false };
 
 const createWrapper = props => shallowMount(DocumentationTopic, {
   stubs,
-  provide,
   mocks,
   ...props,
 });
@@ -196,91 +158,15 @@ describe('DocumentationTopic', () => {
     window.renderedTimes = null;
   });
 
-  it('provides a global store', () => {
-    // eslint-disable-next-line no-underscore-dangle
-    expect(wrapper.vm._provided.store).toEqual(DocumentationTopicStore);
-  });
-
   it('calls the onPageLoadScrollToFragment mixin', () => {
     expect(onPageLoadScrollToFragment.mounted).toHaveBeenCalled();
   });
 
-  it('renders an CodeTheme without `topicData`', () => {
-    wrapper.setData({ topicData: null });
+  it('sets enableNavigator to true if schemaVersion is compatible', async () => {
+    wrapper = createWrapper();
+    wrapper.setData({ topicData });
 
-    const codeTheme = wrapper.find(CodeTheme);
-    expect(codeTheme.exists()).toBe(true);
-    expect(codeTheme.isEmpty()).toEqual(true);
-  });
-
-  it('renders the Navigator and AdjustableSidebarWidth when enabled', async () => {
-    wrapper.setData({
-      topicData: {
-        ...topicData,
-        schemaVersion: schemaVersionWithSidebar,
-      },
-    });
-    const adjustableWidth = wrapper.find(AdjustableSidebarWidth);
-    expect(adjustableWidth.classes())
-      .toEqual(expect.arrayContaining(['full-width-container', 'topic-wrapper']));
-    expect(adjustableWidth.props()).toEqual({
-      shownOnMobile: false,
-      hiddenOnLarge: false,
-      fixedWidth: null,
-    });
-    const technology = topicData.references['topic://foo'];
-    expect(wrapper.find(NavigatorDataProvider).props()).toEqual({
-      interfaceLanguage: Language.swift.key.url,
-      technologyUrl: technology.url,
-      apiChangesVersion: null,
-    });
-    // its rendered by default
-    const navigator = wrapper.find(Navigator);
-    expect(navigator.exists()).toBe(true);
-    expect(navigator.props()).toEqual({
-      errorFetching: false,
-      isFetching: true,
-      // assert we are passing the first set of paths always
-      parentTopicIdentifiers: topicData.hierarchy.paths[0],
-      references: topicData.references,
-      symbolKind: topicData.metadata.symbolKind,
-      scrollLockID: AdjustableSidebarWidth.constants.SCROLL_LOCK_ID,
-      // assert we are passing the default technology, if we dont have the children yet
-      technology,
-      apiChanges: null,
-      flatChildren: [],
-      navigatorReferences: {},
-      renderFilterOnTop: false,
-    });
-    expect(dataUtils.fetchIndexPathsData).toHaveBeenCalledTimes(1);
-    await flushPromises();
-    expect(navigator.props()).toEqual({
-      errorFetching: false,
-      isFetching: false,
-      scrollLockID: AdjustableSidebarWidth.constants.SCROLL_LOCK_ID,
-      renderFilterOnTop: false,
-      parentTopicIdentifiers: topicData.hierarchy.paths[0],
-      references: topicData.references,
-      symbolKind: topicData.metadata.symbolKind,
-      technology: TechnologyWithChildren,
-      apiChanges: null,
-      flatChildren: [],
-      navigatorReferences,
-    });
-    // assert the nav is in wide format
-    const nav = wrapper.find(Nav);
-    expect(nav.props('displaySidenav')).toBe(true);
-  });
-
-  it('renders QuickNavigation if enableQuickNavigation is true', () => {
-    getSetting.mockReturnValueOnce(true);
-    wrapper = createWrapper({
-      stubs: {
-        ...stubs,
-        Nav: DocumentationNav,
-        NavBase,
-      },
-    });
+    expect(wrapper.find(BaseNavigatorView).props('enableNavigator')).toBe(false);
 
     wrapper.setData({
       topicData: {
@@ -289,130 +175,10 @@ describe('DocumentationTopic', () => {
       },
     });
 
-    const quickNavigationModalComponent = wrapper.find(QuickNavigationModal);
-    expect(quickNavigationModalComponent.exists()).toBe(true);
+    expect(wrapper.find(BaseNavigatorView).props('enableNavigator')).toBe(true);
   });
 
-  it('does not render QuickNavigation if enableQuickNavigation is false', () => {
-    wrapper = createWrapper({
-      stubs: {
-        ...stubs,
-        Nav: DocumentationNav,
-        NavBase,
-      },
-    });
-
-    wrapper.setData({
-      topicData: {
-        ...topicData,
-        schemaVersion: schemaVersionWithSidebar,
-      },
-    });
-
-    const quickNavigationModalComponent = wrapper.find(QuickNavigationModal);
-    expect(quickNavigationModalComponent.exists()).toBe(false);
-  });
-
-  it('does not render QuickNavigation and MagnifierIcon if enableNavigation is false', () => {
-    getSetting.mockReturnValueOnce(true);
-    wrapper = createWrapper({
-      stubs: {
-        ...stubs,
-        Nav: DocumentationNav,
-        NavBase,
-      },
-    });
-
-    const quickNavigationModalComponent = wrapper.find(QuickNavigationModal);
-    expect(quickNavigationModalComponent.exists()).toBe(false);
-  });
-
-  it('does not render QuickNavigation if enableQuickNavigation is true but IDE is being targeted', () => {
-    getSetting.mockReturnValueOnce(true);
-    wrapper = createWrapper({
-      provide: { isTargetIDE: true },
-      stubs: {
-        ...stubs,
-        Nav: DocumentationNav,
-        NavBase,
-      },
-    });
-
-    wrapper.setData({
-      topicData: {
-        ...topicData,
-        schemaVersion: schemaVersionWithSidebar,
-      },
-    });
-
-    const quickNavigationModalComponent = wrapper.find(QuickNavigationModal);
-    expect(quickNavigationModalComponent.exists()).toBe(false);
-  });
-
-  describe('if breakpoint is small', () => {
-    beforeEach(() => {
-      wrapper = createWrapper({
-        stubs: {
-          AdjustableSidebarWidth: AdjustableSidebarWidthSmallStub,
-          NavigatorDataProvider,
-        },
-      });
-    });
-
-    it('applies display none to Navigator if is closed', async () => {
-      // renders a closed navigator
-      wrapper.setData({
-        topicData: {
-          ...topicData,
-          schemaVersion: schemaVersionWithSidebar,
-        },
-      });
-      await wrapper.vm.$nextTick();
-      // assert navigator has display: none
-      expect(wrapper.find(Navigator).attributes('style')).toContain('display: none');
-    });
-
-    it('reverses the filter position of the navigator', async () => {
-      // renders a closed navigator
-      wrapper.setData({
-        topicData: {
-          ...topicData,
-          schemaVersion: schemaVersionWithSidebar,
-        },
-      });
-      await wrapper.vm.$nextTick();
-      // assert navigator has display: none
-      expect(wrapper.find(Navigator).props('renderFilterOnTop')).toBe(true);
-    });
-
-    it('does not apply display none to Navigator if is open', async () => {
-      // renders an open navigator
-      wrapper.setData({
-        topicData: {
-          ...topicData,
-          schemaVersion: schemaVersionWithSidebar,
-        },
-        sidenavVisibleOnMobile: true,
-      });
-      await wrapper.vm.$nextTick();
-      // assert navigator doesn't have display: none
-      expect(wrapper.find(Navigator).attributes('style')).toBeFalsy();
-    });
-  });
-
-  it('provides the selected api changes, to the NavigatorDataProvider', () => {
-    wrapper.vm.store.state.selectedAPIChangesVersion = 'latest_major';
-    wrapper.setData({
-      topicData: {
-        ...topicData,
-        schemaVersion: schemaVersionWithSidebar,
-      },
-    });
-    const dataProvider = wrapper.find(NavigatorDataProvider);
-    expect(dataProvider.props('apiChangesVersion')).toEqual('latest_major');
-  });
-
-  it('renders the Navigator with data when no reference is found for a top-level collection', () => {
+  it('passes a technology to the BaseNavigatorView when no reference is found for a top-level collection', () => {
     const technologies = {
       id: 'topic://technologies',
       title: 'Technologies',
@@ -439,126 +205,13 @@ describe('DocumentationTopic', () => {
       },
     });
 
-    const navigator = wrapper.find(Navigator);
-    expect(navigator.exists()).toBe(true);
+    const baseNavigatorViewComponent = wrapper.find(BaseNavigatorView);
+    expect(baseNavigatorViewComponent.exists()).toBe(true);
     // assert the technology is the last fallback
-    expect(navigator.props('technology')).toEqual({
+    expect(baseNavigatorViewComponent.props('technology')).toEqual({
       title: topicData.metadata.title,
       url: mocks.$route.path,
     });
-  });
-
-  it('renders the Navigator with data when no reference is found for a top-level item', () => {
-    const technologies = {
-      id: 'topic://not-existing',
-      title: 'Technologies',
-      url: '/technologies',
-      kind: 'technologies',
-    };
-    wrapper.setData({
-      topicData: {
-        ...topicData,
-        schemaVersion: schemaVersionWithSidebar,
-        hierarchy: {
-          paths: [
-            [technologies.id, ...topicData.hierarchy.paths[0]],
-          ],
-        },
-      },
-    });
-
-    const navigator = wrapper.find(Navigator);
-    expect(navigator.exists()).toBe(true);
-    // assert the technology is the last fallback
-    expect(navigator.props('technology')).toEqual({
-      title: topicData.metadata.title,
-      url: mocks.$route.path,
-    });
-  });
-
-  it('renders the Navigator with data when no reference is found, for any if if the breadcrumbs', () => {
-    const technologies = {
-      id: 'topic://not-existing',
-      title: 'Technologies',
-      url: '/technologies',
-      kind: 'technologies',
-    };
-    wrapper.setData({
-      topicData: {
-        ...topicData,
-        schemaVersion: schemaVersionWithSidebar,
-        hierarchy: {
-          paths: [
-            [technologies.id, ...topicData.hierarchy.paths[0]],
-          ],
-        },
-        // simulate reference data error
-        references: {},
-      },
-    });
-
-    const navigator = wrapper.find(Navigator);
-    expect(navigator.exists()).toBe(true);
-    // assert the technology is the last fallback
-    expect(navigator.props('technology')).toEqual({
-      title: topicData.metadata.title,
-      url: mocks.$route.path,
-    });
-  });
-
-  it('renders the Navigator with data when no hierarchy and reference is found for the current page', () => {
-    wrapper.setData({
-      topicData: {
-        ...topicData,
-        schemaVersion: schemaVersionWithSidebar,
-        // remove the hierarchy items
-        hierarchy: {},
-        // remove the references as well, so it falls back to the last fallback
-        references: {},
-      },
-    });
-
-    const navigator = wrapper.find(Navigator);
-    expect(navigator.exists()).toBe(true);
-    // assert the technology is the last fallback
-    expect(navigator.props('technology')).toEqual({
-      title: topicData.metadata.title,
-      url: mocks.$route.path,
-    });
-  });
-
-  it('renders without a sidebar', () => {
-    wrapper.setData({ topicData });
-
-    // assert the Nav
-    const nav = wrapper.find(Nav);
-    expect(nav.props()).toEqual({
-      isDark: false,
-      hasNoBorder: false,
-      displaySidenav: false,
-      interfaceLanguage: topicData.identifier.interfaceLanguage,
-      objcPath: topicData.variants[0].paths[0],
-      swiftPath: topicData.variants[1].paths[0],
-    });
-    expect(nav.attributes()).toMatchObject({
-      interfacelanguage: 'swift',
-      objcpath: 'documentation/objc',
-      swiftpath: 'documentation/swift',
-    });
-
-    // assert the sidebar
-    expect(wrapper.find(AdjustableSidebarWidth).exists()).toBe(false);
-    const staticContentWidth = wrapper.find(StaticContentWidth);
-    expect(staticContentWidth.exists()).toBe(true);
-    expect(wrapper.find(Navigator).exists()).toBe(false);
-    // assert the proper container class is applied
-    expect(staticContentWidth.classes())
-      .toEqual(expect.arrayContaining(['topic-wrapper', 'full-width-container']));
-  });
-
-  it('renders without NavigatorDataProvider', async () => {
-    wrapper.setData({ topicData });
-    expect(wrapper.find(NavigatorDataProvider).exists()).toBe(false);
   });
 
   it('finds the parentTopicIdentifiers, that have the closest url structure to the current page', () => {
@@ -571,89 +224,10 @@ describe('DocumentationTopic', () => {
           'topic://baz': { url: '/documentation/somepath' },
           'topic://baq': { url: '/documentation/somepath/page' },
         },
-        schemaVersion: schemaVersionWithSidebar,
       },
     });
-    expect(wrapper.find(Navigator).props('parentTopicIdentifiers'))
+    expect(wrapper.find(BaseNavigatorView).props('parentTopicIdentifiers'))
       .toEqual(topicData.hierarchy.paths[1]);
-  });
-
-  it('handles the `@close`, on Navigator, for Mobile breakpoints', async () => {
-    wrapper.setData({
-      topicData: {
-        ...topicData,
-        schemaVersion: schemaVersionWithSidebar,
-      },
-    });
-    await flushPromises();
-    const navigator = wrapper.find(Navigator);
-    const nav = wrapper.find(Nav);
-    // toggle the navigator from the Nav component, in Small breakpoint
-    nav.vm.$emit('toggle-sidenav', BreakpointName.small);
-    const sidebar = wrapper.find(AdjustableSidebarWidth);
-    // set the breakpoint to small on the sidebar
-    sidebar.vm.breakpoint = BreakpointName.small;
-    expect(sidebar.props('shownOnMobile')).toBe(true);
-    await flushPromises();
-    navigator.vm.$emit('close');
-    expect(sidebar.props('shownOnMobile')).toBe(false);
-    // Test that Medium works with the same set of props/events
-    // toggle the navigator from the Nav component, in Medium breakpoint
-    nav.vm.$emit('toggle-sidenav', BreakpointName.medium);
-    expect(sidebar.props('shownOnMobile')).toBe(true);
-    await flushPromises();
-    sidebar.vm.breakpoint = BreakpointName.medium;
-    navigator.vm.$emit('close');
-    expect(sidebar.props('shownOnMobile')).toBe(false);
-    expect(storage.set).toHaveBeenCalledTimes(0);
-  });
-
-  it('handles the `@close`, on Navigator, for `Large` breakpoints', async () => {
-    wrapper.setData({
-      topicData: {
-        ...topicData,
-        schemaVersion: schemaVersionWithSidebar,
-      },
-    });
-    await flushPromises();
-    const sidebar = wrapper.find(AdjustableSidebarWidth);
-    const nav = wrapper.find(Nav);
-    // close the navigator
-    wrapper.find(Navigator).vm.$emit('close');
-    // assert its closed on Large
-    expect(sidebar.props('hiddenOnLarge')).toBe(true);
-    // now toggle it back from the Nav
-    nav.vm.$emit('toggle-sidenav', BreakpointName.large);
-    await flushPromises();
-    // assert its no longer hidden
-    expect(sidebar.props('hiddenOnLarge')).toBe(false);
-  });
-
-  it('handles `@toggle-sidenav` on Nav, for `Large` breakpoint', async () => {
-    // assert that the storage was called to get the navigator closed state from LS
-    expect(storage.get).toHaveBeenCalledTimes(1);
-    expect(storage.get).toHaveBeenCalledWith(NAVIGATOR_HIDDEN_ON_LARGE_KEY, false);
-
-    wrapper.setData({
-      topicData: {
-        ...topicData,
-        schemaVersion: schemaVersionWithSidebar,
-      },
-    });
-    await flushPromises();
-    const nav = wrapper.find(Nav);
-    const sidebar = wrapper.find(AdjustableSidebarWidth);
-    // assert the hidden prop is false
-    expect(sidebar.props('hiddenOnLarge')).toBe(false);
-    // Now close from the sidebar
-    sidebar.vm.$emit('update:hiddenOnLarge', true);
-    expect(sidebar.props('hiddenOnLarge')).toBe(true);
-    expect(storage.set).toHaveBeenLastCalledWith(NAVIGATOR_HIDDEN_ON_LARGE_KEY, true);
-    // now toggle it back, from within the Nav button
-    nav.vm.$emit('toggle-sidenav', BreakpointName.large);
-    // assert we are storing the updated values
-    expect(sidebar.props('hiddenOnLarge')).toBe(false);
-    expect(storage.set).toHaveBeenLastCalledWith(NAVIGATOR_HIDDEN_ON_LARGE_KEY, false);
   });
 
   it('renders a `Topic` with `topicData`', () => {
@@ -681,13 +255,7 @@ describe('DocumentationTopic', () => {
   });
 
   it('renders an inactive link, when no technologies root paths', () => {
-    wrapper = createWrapper({
-      stubs: {
-        ...stubs,
-        Nav: DocumentationNav,
-        NavBase,
-      },
-    });
+    wrapper = createWrapper();
 
     wrapper.setData({ topicData });
 
@@ -697,14 +265,7 @@ describe('DocumentationTopic', () => {
   });
 
   it('renders the title "Documentation" link, if there is root link', () => {
-    wrapper = createWrapper({
-      stubs: {
-        ...stubs,
-        Nav: DocumentationNav,
-        NavBase,
-        'router-link': RouterLinkStub,
-      },
-    });
+    wrapper = createWrapper();
 
     wrapper.setData({
       topicData: {
@@ -771,11 +332,6 @@ describe('DocumentationTopic', () => {
     getSetting.mockReturnValue(false);
     wrapper = createWrapper({
       provide: { isTargetIDE: true },
-      stubs: {
-        // renders sidebar on a small device
-        AdjustableSidebarWidth: AdjustableSidebarWidthSmallStub,
-        NavigatorDataProvider,
-      },
     });
     wrapper.setData({ topicData });
     expect(wrapper.find(Topic).props('enableOnThisPageNav')).toBe(false);
@@ -990,7 +546,6 @@ describe('DocumentationTopic', () => {
     };
 
     expect(mocks.$bridge.on).toHaveBeenNthCalledWith(1, 'contentUpdate', expect.any(Function));
-    expect(mocks.$bridge.on).toHaveBeenNthCalledWith(2, 'codeColors', expect.any(Function));
     // invoke the callback on the $bridge
     mocks.$bridge.on.mock.calls[0][1](data);
     // assert the data is stored
@@ -1086,22 +641,5 @@ describe('DocumentationTopic', () => {
     expect(dataUtils.fetchDataForRouteEnter).toHaveBeenCalledTimes(1);
     expect(dataUtils.fetchDataForRouteEnter)
       .toHaveBeenCalledWith(params.to, params.from, params.next);
-  });
-
-  describe('isTargetIDE', () => {
-    const provideWithIDETarget = { isTargetIDE: true };
-
-    it('does not render a `Nav`', () => {
-      wrapper = createWrapper({ provide: provideWithIDETarget });
-      wrapper.setData({ topicData });
-      expect(wrapper.contains(Nav)).toBe(false);
-    });
-
-    it('does not render an AdjustableSidebarWidth', () => {
-      wrapper = createWrapper({ provide: provideWithIDETarget });
-      wrapper.setData({ topicData });
-      expect(wrapper.find(AdjustableSidebarWidth).exists()).toBe(false);
-      expect(wrapper.find(Topic).exists()).toBe(true);
-    });
   });
 });
